Fix for gh-386. Close gh-386. Close gh-366.
As indicated in the comments added, there is one issue with the
solution: when members switch shifts with this fix, I had to make it
possible for them to switch with shifts that ARE setup as recurring
shifts. As a consequence, it is possible for the same shift to be filled
by two people in the following scenario:
 - Member A switches their shift (Produce at 12:30pm) two months in the future to June 15th.
 - Member B later is signed up for a recurring produce shift that begins on
   May 18th at 12:30pm and recurs every four weeks

Both members in the above scenario will be signed up for the same shift
on June 15th: Produce at 12:30pm.

Dana and I agreed that this will probably happen only on rare
occurrences, so we are going ahead with this change and will fix it if
it really does appear to be a problem.