Updated job types as follows: - Member is now "Member: Closed" - Added "Member: Open" and "Disabled"
"Member: Open" jobs will be jobs that are available on the workshift
interest form (gh-397). IOW: jobs that are open to members

"Member: Closed" jobs will be jobs that are not usually open to members,
but exist because sometimes we need them: Dancers for example

"Disabled" is for jobs that no longer exist. A staff member can mark a
job as disabled in lieu of deleting it (because we don't delete things!)