Have begun moving member hours balance to member model
Currently hours balance are associated with accounts. The
membership department has requested that we move the hours
balance field to the members model so that hours owed can
be associated with the individual members that owe them.

This update creates a means for the membership department
to begin migrating owed hours over from accounts to members.
When the hours have all been moved over, I will be disabling
the views and reports associated with account hour balances.